# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rick and Morty API",
        "description": "API REST de gestión de personajes y favoritos para la prueba técnica.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registra un nuevo usuario",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Rick Sanchez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "rick@citadel.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123",
                                        "minLength": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario registrado exitosamente."
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|abc123xyz..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación (email duplicado o campos inválidos)"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Inicia sesión y genera un token Sanctum",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|abc123xyz..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciales incorrectas"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cierra la sesión revocando el token actual",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sesión cerrada correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/characters": {
            "get": {
                "tags": [
                    "Personajes"
                ],
                "summary": "Listado de personajes paginado y filtrado",
                "operationId": "ad1aa009bd5672d0869aed6ca623d809",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nombre",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado (Alive, Dead, unknown)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "species",
                        "in": "query",
                        "description": "Filtrar por especie",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de personajes"
                    }
                }
            }
        },
        "/api/characters/{character}": {
            "get": {
                "tags": [
                    "Personajes"
                ],
                "summary": "Obtiene el detalle de un personaje específico",
                "operationId": "8cc252a14cb6b540af7584285a98081f",
                "parameters": [
                    {
                        "name": "character",
                        "in": "path",
                        "description": "ID interno del personaje en la base de datos",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del personaje, incluyendo episodios relacionados"
                    },
                    "404": {
                        "description": "Personaje no encontrado"
                    }
                }
            }
        },
        "/api/characters/{character}/favorite": {
            "post": {
                "tags": [
                    "Favoritos"
                ],
                "summary": "Añade o elimina un personaje de favoritos",
                "operationId": "21a3b3fb54a459aa8a8cd9b703c7877b",
                "parameters": [
                    {
                        "name": "character",
                        "in": "path",
                        "description": "ID del personaje",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado del favorito actualizado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Personaje no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favorites": {
            "get": {
                "tags": [
                    "Favoritos"
                ],
                "summary": "Listado de personajes favoritos del usuario autenticado",
                "operationId": "0914ea375285d2842e43553a8cf63d7c",
                "responses": {
                    "200": {
                        "description": "Lista paginada de los personajes marcados como favoritos"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticación",
            "description": "Autenticación"
        },
        {
            "name": "Personajes",
            "description": "Personajes"
        },
        {
            "name": "Favoritos",
            "description": "Favoritos"
        }
    ]
}